[dynamo][guards] Get rid of FUNCTION_MATCH
In the previous PRs, we have gotten rid of all FUNCTION_MATCH, but I am
hesitant to entirely remove the ID_MATCH because it can cause a failure
in the wild. So, replacing with UNCLASSIFIED_ID_MATCH. This will show up
in the recompiles reasons in the tlparse if we overguard, and therefore
will be easier to track.

ghstack-source-id: 1909ddc66923934ec85ebcd08df9f8d3740d7155
Pull Request resolved: https://github.com/pytorch/pytorch/pull/166321