Enable partially local tensor mode for DTensor attention and convolution tests
Attention implementation uses rank local logic (computation based on local rank)
that is not compatible with local tensor out of the box. In order to enale running
attention under local tensor we will need to rewrite parts of it to make it local
tensor friendly. This change fixes several issues in the local tensor implementation
that are prerequisite for attention to be runnable, including proper handling of
the inplace ops, tensor comparison and flattening/unflattening.

ghstack-source-id: aced3aa999bd4b17186b7fa9c1b175dd96a9436d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/166406