[CP][BE][2/2] Refactor the code structure
Our CP codebase now contains several files and we are adding more. This
PR refactors the code to consolidate the files into a context_parallel
folder but keep the import so that the existing users of CP won't be
affected.

Unfortunately, we have to split this PR into two PRs as the PyTorch
infra cannot accept a PR with 3000+ LoC change and git cannot recognize
that _context_parallel/_attention.py is moved from _attention.py because
we want to keep BC.

This is the second PR.

ghstack-source-id: e792fbaeedfb927dc245ae008ff0ce57ef2d13a1
Pull-Request: https://github.com/pytorch/pytorch/pull/166501